+++ packages/browser-wallet/src/popup/popupX/pages/ConnectedSites/ConnectedSites.tsx
@@ -1,50 +1,40 @@
 import React, { useMemo } from 'react';
-import Page from '@popup/popupX/shared/Page';
+import { useAtom } from 'jotai';
+import { generatePath, useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
+import Page from '@popup/popupX/shared/Page';
 import Text from '@popup/popupX/shared/Text';
 import Card from '@popup/popupX/shared/Card';
 import Button from '@popup/popupX/shared/Button';
-import { Navigate, useParams } from 'react-router-dom';
-import { useAtom } from 'jotai';
 import { storedAllowlistAtom } from '@popup/store/account';
-import { displayNameAndSplitAddress, useCredential } from '@popup/shared/utils/account-helpers';
-import { popupMessageHandler } from '@popup/shared/message-handler';
-import { EventType } from '@concordium/browser-wallet-api-helpers';
+import { displayUrl } from '@popup/shared/utils/string-helpers';
+import { absoluteRoutes } from '@popup/popupX/constants/routes';
 
-type ConnectedSitesProps = {
-    address: string;
-};
-
-function ConnectedSites({ address }: ConnectedSitesProps) {
+function ConnectedSites() {
     const { t } = useTranslation('x', { keyPrefix: 'connectedSites' });
-    const [allowlistWithLoading, setAllowList] = useAtom(storedAllowlistAtom);
-    const credential = useCredential(address);
+    const nav = useNavigate();
+    const [allowlistWithLoading] = useAtom(storedAllowlistAtom);
     const allowlist = allowlistWithLoading.value ?? {};
-    const connectedAccountSites = useMemo(
-        () =>
-            Object.entries(allowlist).flatMap(([url, accounts]) => {
-                const accountIndex = accounts.findIndex((a) => a === address);
-                return accountIndex === -1 ? [] : [{ url, accountIndex }];
-            }),
-        [allowlist, address]
-    );
-    const onDisconnect = (url: string, accountIndex: number) => async () => {
-        const updatedAllowlist = { ...allowlist, [url]: allowlist[url].toSpliced(accountIndex, 1) };
-        await setAllowList(updatedAllowlist);
-        popupMessageHandler.broadcastToUrl(EventType.AccountDisconnected, url, address);
-    };
+    const connectedSites = useMemo(() => Object.keys(allowlist), [allowlist]);
+    const navToEdit = (serviceName: string) =>
+        nav(
+            generatePath(absoluteRoutes.settings.accounts.connectedSites.edit.path, {
+                serviceName: encodeURIComponent(serviceName),
+            })
+        );
     const loadedSiteView =
-        connectedAccountSites.length === 0 ? (
-            <Text.MainRegular>No sites connected to this account</Text.MainRegular>
+        connectedSites.length === 0 ? (
+            <Text.MainRegular>{t('noConnectedSites')}</Text.MainRegular>
         ) : (
             <Card>
-                {connectedAccountSites.map((site) => (
-                    <Card.Row key={site.url}>
-                        <Text.MainRegular>{new URL(site.url).host}</Text.MainRegular>
-                        <Button.Secondary
-                            className="dark"
-                            label={t('disconnect')}
-                            onClick={onDisconnect(site.url, site.accountIndex)}
+                {connectedSites.map((serviceName) => (
+                    <Card.Row key={serviceName}>
+                        <Text.MainRegular>{displayUrl(serviceName)}</Text.MainRegular>
+                        <Button.Main
+                            size="small"
+                            variant="secondary"
+                            label={t('edit')}
+                            onClick={() => navToEdit(serviceName)}
                         />
                     </Card.Row>
                 ))}
@@ -52,19 +42,10 @@ function ConnectedSites({ address }: ConnectedSitesProps) {
         );
     return (
         <Page className="connected-sites-x">
-            <Page.Top heading={t('connectedSites')}>
-                <Text.Capture>{displayNameAndSplitAddress(credential)}</Text.Capture>
-            </Page.Top>
+            <Page.Top heading={t('connectedSites')} />
             <Page.Main>{allowlistWithLoading.loading ? null : loadedSiteView}</Page.Main>
         </Page>
     );
 }
 
-export default function Loader() {
-    const params = useParams();
-    if (params.account === undefined) {
-        // No account address passed in the url.
-        return <Navigate to="../" />;
-    }
-    return <ConnectedSites address={params.account} />;
-}
+export default ConnectedSites;