+++ packages/browser-wallet/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 -   Added support for two additional types of Sponsored Transactions, `Configure Delegation` and `Update Smart Contract Instance`.
+-   Added Allowlist config at account page as part of feature parity with Old UI
+-   Added export activity button to Transaction Log as part of feature parity with Old UI
 
 ## 2.9.0